requirements.txt
Flask>=2.3.0,<3.0.0
Flask-CORS>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
ultralytics>=8.0.0,<9.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
gunicorn>=21.0.0 